move hooks to own submodule
